#nest-v4 : move path solver into its own file
