+++ test/utils/path.solver.test.ts
@@ -1,15 +1,6 @@
-import { basename, dirname, normalize, Path, relative } from '@angular-devkit/core';
+import { normalize, Path } from '@angular-devkit/core';
 import { expect } from 'chai';
-
-export class PathSolver {
-  constructor() {}
-
-  public relative(from: Path, to: Path): string {
-    const relativeDir: Path = relative(dirname(from), dirname(to));
-    return (relativeDir.startsWith('.') ? relativeDir : './' + relativeDir)
-      .concat(relativeDir.length === 0 ? basename(to) : '/' + basename(to));
-  }
-}
+import { PathSolver } from '../../src/utils/path.solver';
 
 describe('Path Solver', () => {
   it('should relative path between path', () => {
